+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **Breaking:** `useLocalize` is removed. API localizes configs.
 - **Deprecation:** `field.name` is deprecated. Use `field.title` instead.
+- Added `onPostSubmit` prop to get notified when form has been submitted
 
 # 0.1.0